fix: Increase test coverage.
