+++ optimizely/client/factory_test.go
@@ -23,6 +23,7 @@ import (
 	"github.com/optimizely/go-sdk/optimizely/config"
 	"github.com/optimizely/go-sdk/optimizely/config/datafileprojectconfig"
 	"github.com/optimizely/go-sdk/optimizely/event"
+	"github.com/optimizely/go-sdk/optimizely/utils"
 
 	"github.com/stretchr/testify/assert"
 )
@@ -84,4 +85,25 @@ func TestClientWithDecisionServiceAndEventProcessorInOptions(t *testing.T) {
 	assert.NoError(t, err)
 	assert.Equal(t, decisionService, optimizelyClient.DecisionService)
 	assert.Equal(t, processor, optimizelyClient.EventProcessor)
+}
+
+func TestClientWithCustomCtx(t *testing.T) {
+	factory := OptimizelyFactory{}
+	testExecutionCtx := utils.NewCancelableExecutionCtx()
+	mockConfigManager := new(MockProjectConfigManager)
+	client, err := factory.Client(
+		WithConfigManager(mockConfigManager),
+		WithExecutionContext(testExecutionCtx),
+	)
+	assert.NoError(t, err)
+	assert.Equal(t, client.executionCtx, testExecutionCtx)
+}
+
+func TestStaticClient(t *testing.T) {
+	factory := OptimizelyFactory{Datafile: []byte(`{"revision": "42"}`)}
+	optlyClient, err := factory.StaticClient()
+	assert.NoError(t, err)
+
+	parsedConfig, _ := optlyClient.ConfigManager.GetConfig()
+	assert.Equal(t, "42", parsedConfig.GetRevision())
 }
